Separated media and trending routes
