+++ index.js
@@ -3,6 +3,7 @@ import 'dotenv/config'
 import cors from 'cors';
 import { router as mediaRoute } from "./routes/mediaRoute.js"
 import { router as userRoute } from './routes/userRoute.js'
+import { router as trendingRoute } from './routes/trendingRoute.js'
 
 const app = express()
 const PORT = process.env.PORT || 10000
@@ -17,6 +18,7 @@ app.get('/', (req, res) => {
 
 app.use("/api/media", mediaRoute)
 app.use("/api/users", userRoute)
+app.use("/api/trending", trendingRoute)
 
 app.listen(PORT, () => {
     console.log(`This server is running in port ${PORT}`)
